Refactor CommunityFragment by creating PostFragment, Disable pet profile temporarily
